+++ databento/live/gateway.py
@@ -7,10 +7,12 @@ from operator import attrgetter
 from typing import SupportsBytes
 from typing import TypeVar
 
+from databento_dbn import Compression
 from databento_dbn import Encoding
 from databento_dbn import Schema
 from databento_dbn import SType
 
+from databento.common.enums import SlowReaderBehavior
 from databento.common.publishers import Dataset
 from databento.common.system import USER_AGENT
 
@@ -117,8 +119,15 @@ class AuthenticationRequest(GatewayControl):
     encoding: Encoding = Encoding.DBN
     details: str | None = None
     ts_out: str = "0"
+    compression: Compression | str = Compression.NONE
     heartbeat_interval_s: int | None = None
+    slow_reader_behavior: SlowReaderBehavior | str | None = None
     client: str = USER_AGENT
+
+    def __post_init__(self) -> None:
+        # Temporary work around for LSG support
+        if self.slow_reader_behavior in [SlowReaderBehavior.SKIP, "skip"]:
+            self.slow_reader_behavior = "drop"
 
 
 @dataclasses.dataclass