+++ databento/live/session.py
@@ -14,12 +14,14 @@ from typing import Final
 
 import databento_dbn
 import pandas as pd
+from databento_dbn import Compression
 from databento_dbn import DBNRecord
 from databento_dbn import Schema
 from databento_dbn import SType
 
 from databento.common.constants import ALL_SYMBOLS
 from databento.common.enums import ReconnectPolicy
+from databento.common.enums import SlowReaderBehavior
 from databento.common.error import BentoError
 from databento.common.publishers import Dataset
 from databento.common.types import ClientRecordCallback
@@ -205,8 +207,17 @@ class _SessionProtocol(DatabentoLiveProtocol):
         metadata: SessionMetadata,
         ts_out: bool = False,
         heartbeat_interval_s: int | None = None,
+        slow_reader_behavior: SlowReaderBehavior | str | None = None,
+        compression: Compression = Compression.NONE,
     ):
-        super().__init__(api_key, dataset, ts_out, heartbeat_interval_s)
+        super().__init__(
+            api_key,
+            dataset,
+            ts_out,
+            heartbeat_interval_s,
+            slow_reader_behavior,
+            compression,
+        )
 
         self._dbn_queue = dbn_queue
         self._loop = loop
@@ -302,6 +313,8 @@ class LiveSession:
         The reconnect policy for the live session.
             - "none": the client will not reconnect (default)
             - "reconnect": the client will reconnect automatically
+    compression : Compression, optional
+        The compression format for the session. Defaults to no compression.
     """
 
     def __init__(
@@ -313,6 +326,8 @@ class LiveSession:
         user_gateway: str | None = None,
         user_port: int = DEFAULT_REMOTE_PORT,
         reconnect_policy: ReconnectPolicy | str = ReconnectPolicy.NONE,
+        slow_reader_behavior: SlowReaderBehavior | str | None = None,
+        compression: Compression = Compression.NONE,
     ) -> None:
         self._dbn_queue = DBNQueue()
         self._lock = threading.RLock()
@@ -329,6 +344,8 @@ class LiveSession:
         self._api_key = api_key
         self._ts_out = ts_out
         self._heartbeat_interval_s = heartbeat_interval_s or 30
+        self._slow_reader_behavior = slow_reader_behavior
+        self._compression = compression
 
         self._protocol: _SessionProtocol | None = None
         self._transport: asyncio.Transport | None = None
@@ -579,6 +596,8 @@ class LiveSession:
             metadata=self._metadata,
             ts_out=self.ts_out,
             heartbeat_interval_s=self.heartbeat_interval_s,
+            slow_reader_behavior=self._slow_reader_behavior,
+            compression=self._compression,
         )
 
     def _connect(